requirement.txt
numpy
pandas
matplotlib
seaborn
tensorflow
keras
opencv-python
pillow
flask
